+++ common/src/main/java/smartin/miapi/client/model/module/BakedMiapiModel.java
@@ -38,7 +38,7 @@ import smartin.miapi.client.renderer.TrimRenderer;
 import smartin.miapi.config.MiapiConfig;
 import smartin.miapi.item.modular.Transform;
 import smartin.miapi.material.MaterialProperty;
-import smartin.miapi.mixin.BufferBuilderAccessor;
+import smartin.miapi.mixin.client.BufferBuilderAccessor;
 import smartin.miapi.modules.ModuleInstance;
 import smartin.miapi.modules.properties.GlintProperty;
 import smartin.miapi.modules.properties.render.AlphaOverwriteProperty;