+++ fabric/src/main/java/smartin/miapi/fabric/MiapiFabric.java
@@ -21,7 +21,7 @@ import smartin.miapi.client.MiapiClient;
 import smartin.miapi.datapack.ReloadEvents;
 import smartin.miapi.events.MiapiEvents;
 import smartin.miapi.item.modular.ModularItem;
-import smartin.miapi.mixin.OptionsAccessor;
+import smartin.miapi.mixin.client.OptionsAccessor;
 import smartin.miapi.mixin.client.KeyMappingAccessor;
 import smartin.miapi.modules.properties.attributes.AttributeProperty;
 import smartin.miapi.modules.properties.enchanment.AllowedEnchantments;